refactor(complete): Organize around completion UI
