--- clap_complete/src/dynamic/shells/elvish.rs
@@ -1,59 +1,0 @@
-/// A [`ShellCompleter`][crate::dynamic::shells::ShellCompleter] for Elvish
-#[derive(Copy, Clone, PartialEq, Eq, Debug)]
-pub struct Elvish;
-
-impl crate::dynamic::shells::ShellCompleter for Elvish {
-    fn file_name(&self, name: &str) -> String {
-        format!("{name}.elv")
-    }
-    fn write_registration(
-        &self,
-        _name: &str,
-        bin: &str,
-        completer: &str,
-        buf: &mut dyn std::io::Write,
-    ) -> Result<(), std::io::Error> {
-        let bin = shlex::try_quote(bin).unwrap_or(std::borrow::Cow::Borrowed(bin));
-        let completer =
-            shlex::try_quote(completer).unwrap_or(std::borrow::Cow::Borrowed(completer));
-
-        let script = r#"
-set edit:completion:arg-completer[BIN] = { |@words|
-    set E:_CLAP_IFS = "\n"
-
-    var index = (count $words)
-    set index = (- $index 1)
-    set E:_CLAP_COMPLETE_INDEX = (to-string $index)
-
-    put (COMPLETER complete --shell elvish -- $@words) | to-lines
-}
-"#
-        .replace("COMPLETER", &completer)
-        .replace("BIN", &bin);
-
-        writeln!(buf, "{script}")?;
-        Ok(())
-    }
-    fn write_complete(
-        &self,
-        cmd: &mut clap::Command,
-        args: Vec<std::ffi::OsString>,
-        current_dir: Option<&std::path::Path>,
-        buf: &mut dyn std::io::Write,
-    ) -> Result<(), std::io::Error> {
-        let index: usize = std::env::var("_CLAP_COMPLETE_INDEX")
-            .ok()
-            .and_then(|i| i.parse().ok())
-            .unwrap_or_default();
-        let ifs: Option<String> = std::env::var("_CLAP_IFS").ok().and_then(|i| i.parse().ok());
-        let completions = crate::dynamic::complete(cmd, args, index, current_dir)?;
-
-        for (i, candidate) in completions.iter().enumerate() {
-            if i != 0 {
-                write!(buf, "{}", ifs.as_deref().unwrap_or("\n"))?;
-            }
-            write!(buf, "{}", candidate.get_content().to_string_lossy())?;
-        }
-        Ok(())
-    }
-}
