--- clap_complete/src/dynamic/shells/shell.rs
@@ -1,166 +1,0 @@
-use std::fmt::Display;
-use std::str::FromStr;
-
-use clap::builder::PossibleValue;
-use clap::ValueEnum;
-
-/// Completion support for built-in shells
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-#[non_exhaustive]
-pub enum Shell {
-    /// Bourne Again `SHell` (bash)
-    Bash,
-    /// Elvish shell
-    Elvish,
-    /// Friendly Interactive `SHell` (fish)
-    Fish,
-    /// `PowerShell`
-    Powershell,
-    /// Z `SHell` (zsh)
-    Zsh,
-}
-
-impl Shell {
-    /// Parse a shell from a path to the executable for the shell
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use clap_complete::shells::Shell;
-    ///
-    /// assert_eq!(Shell::from_shell_path("/bin/bash"), Some(Shell::Bash));
-    /// assert_eq!(Shell::from_shell_path("/usr/bin/zsh"), Some(Shell::Zsh));
-    /// assert_eq!(Shell::from_shell_path("/opt/my_custom_shell"), None);
-    /// ```
-    pub fn from_shell_path(path: impl AsRef<std::path::Path>) -> Option<Shell> {
-        parse_shell_from_path(path.as_ref())
-    }
-
-    /// Determine the user's current shell from the environment
-    ///
-    /// This will read the SHELL environment variable and try to determine which shell is in use
-    /// from that.
-    ///
-    /// If SHELL is not set, then on windows, it will default to powershell, and on
-    /// other operating systems it will return `None`.
-    ///
-    /// If SHELL is set, but contains a value that doesn't correspond to one of the supported shell
-    /// types, then return `None`.
-    ///
-    /// # Example:
-    ///
-    /// ```no_run
-    /// # use clap::Command;
-    /// use clap_complete::{generate, shells::Shell};
-    /// # fn build_cli() -> Command {
-    /// #     Command::new("compl")
-    /// # }
-    /// let shell = Shell::from_env();
-    /// println!("{shell:?}");
-    /// ```
-    pub fn from_env() -> Option<Shell> {
-        if let Some(env_shell) = std::env::var_os("SHELL") {
-            Shell::from_shell_path(env_shell)
-        } else if cfg!(windows) {
-            Some(Shell::Powershell)
-        } else {
-            None
-        }
-    }
-}
-
-// use a separate function to avoid having to monomorphize the entire function due
-// to from_shell_path being generic
-fn parse_shell_from_path(path: &std::path::Path) -> Option<Shell> {
-    let name = path.file_stem()?.to_str()?;
-    match name {
-        "bash" => Some(Shell::Bash),
-        "elvish" => Some(Shell::Elvish),
-        "fish" => Some(Shell::Fish),
-        "powershell" | "powershell_ise" => Some(Shell::Powershell),
-        "zsh" => Some(Shell::Zsh),
-        _ => None,
-    }
-}
-
-impl Display for Shell {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.to_possible_value()
-            .expect("no values are skipped")
-            .get_name()
-            .fmt(f)
-    }
-}
-
-impl FromStr for Shell {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        for variant in Self::value_variants() {
-            if variant.to_possible_value().unwrap().matches(s, false) {
-                return Ok(*variant);
-            }
-        }
-        Err(format!("invalid variant: {s}"))
-    }
-}
-
-// Hand-rolled so it can work even when `derive` feature is disabled
-impl ValueEnum for Shell {
-    fn value_variants<'a>() -> &'a [Self] {
-        &[
-            Shell::Bash,
-            Shell::Elvish,
-            Shell::Fish,
-            Shell::Powershell,
-            Shell::Zsh,
-        ]
-    }
-
-    fn to_possible_value(&self) -> Option<PossibleValue> {
-        Some(match self {
-            Shell::Bash => PossibleValue::new("bash"),
-            Shell::Elvish => PossibleValue::new("elvish"),
-            Shell::Fish => PossibleValue::new("fish"),
-            Shell::Powershell => PossibleValue::new("powershell"),
-            Shell::Zsh => PossibleValue::new("zsh"),
-        })
-    }
-}
-
-impl Shell {
-    fn completer(&self) -> &dyn crate::dynamic::shells::ShellCompleter {
-        match self {
-            Self::Bash => &super::Bash,
-            Self::Elvish => &super::Elvish,
-            Self::Fish => &super::Fish,
-            Self::Powershell => &super::Powershell,
-            Self::Zsh => &super::Zsh,
-        }
-    }
-}
-
-impl crate::dynamic::shells::ShellCompleter for Shell {
-    fn file_name(&self, name: &str) -> String {
-        self.completer().file_name(name)
-    }
-    fn write_registration(
-        &self,
-        name: &str,
-        bin: &str,
-        completer: &str,
-        buf: &mut dyn std::io::Write,
-    ) -> Result<(), std::io::Error> {
-        self.completer()
-            .write_registration(name, bin, completer, buf)
-    }
-    fn write_complete(
-        &self,
-        cmd: &mut clap::Command,
-        args: Vec<std::ffi::OsString>,
-        current_dir: Option<&std::path::Path>,
-        buf: &mut dyn std::io::Write,
-    ) -> Result<(), std::io::Error> {
-        self.completer().write_complete(cmd, args, current_dir, buf)
-    }
-}
